requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.2
opencv-python-headless==4.8.0.74
mediapipe==0.10.0
numpy==1.24.2
python-socketio==5.8.0
Pillow==9.3.0
eventlet==0.33.3
